# openapi.yaml
openapi: 3.0.3
info:
  title: TiffinBook API
  description: >
    Subscription management and pro-rated billing for home-style tiffin
    (lunch delivery) services. All `/api/customers/*` routes are scoped to
    the logged-in owner and require a Bearer token from register/login.
  version: 1.0.0
servers:
  - url: http://localhost:3000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Customer:
      type: object
      properties:
        id: { type: integer, example: 1 }
        owner_id: { type: integer, example: 1 }
        name: { type: string, example: "Riya Sharma" }
        phone: { type: string, example: "9876543210" }
        plan_price: { type: number, example: 3000 }
        status: { type: string, enum: [active, paused], example: active }
        created_at: { type: string, format: date-time }
        cycle_start: { type: string, format: date }
        transfer_to_customer_id: { type: integer, nullable: true }
...
paths:
...
  /api/customers:
    post:
      summary: Subscribe a new customer
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, phone, planPrice]
              properties:
                name: { type: string }
                phone: { type: string }
                planPrice: { type: number }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Customer' } } } }
        '400': { description: Validation error }
        '409': { description: Phone number already in use by this owner }
    get:
      summary: List customers (search, filter, sort, paginate)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: search
          schema: { type: string }
          description: Matches name or phone (partial, case-insensitive)
        - in: query
          name: status
          schema: { type: string, enum: [active, paused] }
        - in: query
          name: sortBy
          schema: { type: string, enum: [name, phone, plan_price, status, created_at] }
        - in: query
          name: order
          schema: { type: string, enum: [asc, desc] }
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: pageSize
          schema: { type: integer, default: 10 }
      responses:
        '200': { description: Paginated customer list }
...
  /api/notifications/clock:
    post:
      summary: Run the morning delivery notification clock
      security: [{ bearerAuth: [] }]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                date: { type: string, format: date }
      responses:
        '200': { description: Queued delivery reminders for active, non-paused customers on weekdays }
  /api/notifications/outbox:
    get:
      summary: Read the notification outbox
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: date
          schema: { type: string, format: date }
      responses:
        '200': { description: Durable queued notification messages }
  /clock:
    post:
      summary: Assessment alias for the notification clock
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Notification clock result }
  /outbox:
    get:
      summary: Assessment alias for the notification outbox
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Notification outbox rows }
...